requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
netmiko==4.3.0
pydantic-settings==2.2.1
python-multipart==0.0.9
cryptography==41.0.3
pytest==7.3.1
pytest-asyncio==0.20.3
httpx==0.34.0
pytest-cov==4.0.0
coverage==7.2.7
pytest-mock==3.10.0
pytest-cov==4.0.0
pytest-asyncio==0.20.3
